Add an abstracted photo similarity function that can run multiple different analyses
