+++ backend/app/analysis/resnet18_cosine_similarity.py
@@ -1,56 +1,27 @@
 """
-resnet18_feature_vectors.py
+resnet18_cosine_similarity.py
 """
-import sys
-from pathlib import Path
-
 import torch
 from torch import nn
 
-from django.conf import settings
-
 from app.models import Photo
+from backend.app.analysis.photo_similarity import analyze_similarity
 
 MODEL = Photo
-
-
-def deserialize_tensor(photo):
-    dir_path = Path(settings.ANALYSIS_PICKLE_PATH,
-                    'resnet18_features',
-                    str(photo.map_square.number))
-    serialized_feature_vector_path = Path(dir_path, f'{photo.number}.pt')
-    if not serialized_feature_vector_path.exists():
-        print(
-            f'A feature vector for photo {photo.number} in map square {photo.map_square.number} '
-            'was never serialized.'
-            '\nPlease run resnet18_feature_vectors first.\n'
-        )
-        return None
-
-    tensor = torch.load(serialized_feature_vector_path)
-    return tensor
 
 
 def analyze(photo: Photo):
     """
     Produce a list of all other photos by cosine similarity to this photo's feature vector
     """
-    photo_features = deserialize_tensor(photo)
+    return analyze_similarity(photo, cosine_similarity)
 
-    similarities = []
-    for other_photo in Photo.objects.all():
-        other_photo_features = deserialize_tensor(other_photo)
-        if other_photo_features is None:
-            continue
 
-        cosine_similarity_func = nn.CosineSimilarity(dim=1)
-        cosine_similarity = cosine_similarity_func(photo_features, other_photo_features)
-        cosine_similarity_mean = torch.mean(cosine_similarity).item()
-
-        similarities.append(
-            (other_photo.map_square.number, other_photo.number, cosine_similarity_mean)
-        )
-
-    similarities.sort(key=lambda x: x[2])  # sort by cosine_similarity,
-    similarities.reverse()
-    return similarities
+def cosine_similarity(photo_features, other_photo_features):
+    """
+    Compute the cosine similarity between two feature vectors.
+    """
+    cosine_similarity_func = nn.CosineSimilarity(dim=1)
+    similarity = cosine_similarity_func(photo_features, other_photo_features)
+    cosine_similarity_mean = torch.mean(similarity).item()
+    return cosine_similarity_mean
